consolidated SelectTokenButton
